% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos-tools.R
\name{GetOrganizationsFromAllPapers}
\alias{GetOrganizationsFromAllPapers}
\title{Get Organizations From All Papers}
\usage{
GetOrganizationsFromAllPapers(allpapers_path)
}
\arguments{
\item{allpapers_path}{Path to the JSON file containing information about papers.}
}
\value{
A data frame with columns including 'uid', 'ad', and 'oo' representing paper UID,
  organization's full address, and organization's content, respectively.
}
\description{
This function reads JSON data from a file path containing information about papers and extracts
organizations from authors' information.
}
\examples{
allpapers_path <- "path/to/allpapers.json"
organizations_data <- GetOrganizationsFromAllPapers(allpapers_path)

}
