% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos-tools.R
\name{GetOrganizationsFromRawJson}
\alias{GetOrganizationsFromRawJson}
\title{Get organizations data from raw JSON file.}
\usage{
GetOrganizationsFromRawJson(rawJson_path)
}
\arguments{
\item{rawJson_path}{Path to the raw JSON file.}
}
\value{
A data frame containing UID, address (ad), and organization (oo) information.
}
\description{
This function reads a raw JSON file and extracts organizations data.
}
\examples{
# Example usage:
organizations_data <- GetOrganizationsFromRawJson("/path/to/raw_json_file.json")

}
\seealso{
\code{\link{GetAddressesFromRawJson}}
}
